Moved mutableConnection manager under helpers
